+++ tests/test_measurements.py
@@ -23,7 +23,6 @@ import pennylane as qml
 import pytest
 from conftest import LightningDevice as ld
 from conftest import LightningException, device_name, lightning_ops, validate_measurements
-from flaky import flaky
 from pennylane.exceptions import DeviceError, QuantumFunctionError
 from pennylane.measurements import ExpectationMP, Shots, VarianceMP
 from pennylane.wires import Wires